Code clean up to reduce number of warnings from gnu compiler Improve code for ternary alloys computation Added test for computing equilibrium compositions of ternary alloy with database for MoNbTa

+++ source/CALPHADMobility.C
@@ -70,7 +70,7 @@ void CALPHADMobility::initialize(
                                        it++ )
    {
       if( db->keyExists( *it ) ){
-         int n=db->getArraySize(*it);
+         size_t n=db->getArraySize(*it);
          assert(n>=2);
          assert(n<=4);
